refactor(lex): Pull out token mod
